pyproject.toml
[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.mypy]
# check_untyped_defs = true
# disallow_untyped_defs = false
disallow_any_generics = false
disallow_incomplete_defs = false
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true

[tool.uv]
dev-dependencies = ["ipdb>=0.13.13", "ipython>=8.12.3"]

[tool.uv.sources]

[project]
requires-python = ">=3.9"
name = "mtv-api-tests"
version = "0.1.0"
description = "MTV API Tests"
authors = []
license = "Apache-2.0"
readme = "README.md"
dependencies = [
  "click>=8.1.7",
  "humanfriendly>=10.0",
  "openshift>=0.13.2",
  "openshift-python-wrapper>=11.0.13",
  "openstacksdk>=4.0.0",
  "ovirt-python-sdk>=1.2.0",
  "pytest>=8.3.3",
  "pytest-check>=2.4.1",
  "pytest-progress>=1.3.0",
  "pytest-testconfig>=0.2.0",
  "python-glanceclient>=4.7.0",
  "pytz>=2024.2",
  "pyvmomi>=8.0.3.0.1",
  "python-simple-logger>=2.0.0",
  "shortuuid>=1.0.13",
  "openshift-python-wrapper>=11.0.12",
]
